# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for managing categories and products with full CRUD operations\n\n## Features:\n- Category Management (Get all, Get by ID, Create, Update, Delete)\n- Product Management (Get all with category names, Get by ID with category, Create, Update, Delete)\n- Product Search by Name (case-insensitive partial match)\n- Product-Category Relationship (Foreign key with JOIN operations)\n- Transaction / Checkout (Process multi-item checkout with stock deduction)\n- Sales Reports (Daily summary and date range reports with best selling product)\n\n## Response Format:\nAll endpoints return a standard response with:\n- status (bool): Request success status\n- message (string): Response message\n- data (object): Response data (when applicable)",
        "title": "Category Management API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/checkout": {
            "post": {
                "description": "Process a checkout with multiple items. Validates product availability, deducts stock, and creates a transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Process checkout",
                "parameters": [
                    {
                        "description": "Checkout request with items and quantities",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Checkout successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Server error or insufficient stock",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/report": {
            "get": {
                "description": "Get sales summary for a specific date range including total revenue, transaction count, and best selling product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get sales report by date range",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2026-01-01\"",
                        "description": "Start date (YYYY-MM-DD format)",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"2026-02-01\"",
                        "description": "End date (YYYY-MM-DD format)",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales report retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SalesReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Missing required query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to get report",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/report/today": {
            "get": {
                "description": "Get sales summary for today including total revenue, transaction count, and best selling product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get today's sales report",
                "responses": {
                    "200": {
                        "description": "Daily sales report retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.SalesReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to get daily report",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "description": "Retrieve a list of all categories",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all categories",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Category"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new category to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "parameters": [
                    {
                        "description": "Category object that needs to be added",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CategoryInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Category created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Category"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "description": "Retrieve a list of all products with their category names. Supports optional search by name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search product by name (case-insensitive partial match)",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all products",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Product"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new product to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product object that needs to be added",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Product"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.BestSellingProduct": {
            "description": "Best selling product information",
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Indomie Goreng"
                },
                "qty_sold": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
        "models.Category": {
            "description": "Category information with ID, name and description",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-30T12:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Electronic devices and gadgets"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Electronics"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-30T12:00:00Z"
                }
            }
        },
        "models.CategoryInput": {
            "description": "Input model for creating or updating a category (ID is auto-generated)",
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Electronic devices and gadgets"
                },
                "name": {
                    "type": "string",
                    "example": "Electronics"
                }
            }
        },
        "models.CheckoutItem": {
            "description": "Single item to be checked out",
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "integer",
                    "example": 3
                },
                "quantity": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.CheckoutRequest": {
            "description": "Request body for processing a checkout",
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CheckoutItem"
                    }
                }
            }
        },
        "models.Product": {
            "description": "Product information with ID, name, price, stock, and category relationship",
            "type": "object",
            "required": [
                "name",
                "price",
                "stock"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "category_name": {
                    "type": "string",
                    "example": "Electronics"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-30T12:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "iPhone 15 Pro"
                },
                "price": {
                    "type": "integer",
                    "example": 15000000
                },
                "stock": {
                    "type": "integer",
                    "example": 50
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-30T12:00:00Z"
                }
            }
        },
        "models.ProductInput": {
            "description": "Input model for creating or updating a product (ID is auto-generated)",
            "type": "object",
            "required": [
                "name",
                "price",
                "stock"
            ],
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "iPhone 15 Pro"
                },
                "price": {
                    "type": "integer",
                    "example": 15000000
                },
                "stock": {
                    "type": "integer",
                    "example": 50
                }
            }
        },
        "models.Response": {
            "description": "Standard API response structure",
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SalesReport": {
            "description": "Sales summary report with revenue, transaction count, and best seller",
            "type": "object",
            "properties": {
                "best_selling_product": {
                    "$ref": "#/definitions/models.BestSellingProduct"
                },
                "total_revenue": {
                    "type": "integer",
                    "example": 45000
                },
                "total_transactions": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.Transaction": {
            "description": "Transaction information with details of purchased items",
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-02-08T12:00:00Z"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionDetail"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "total_amount": {
                    "type": "integer",
                    "example": 45000
                }
            }
        },
        "models.TransactionDetail": {
            "description": "Detail of a single item within a transaction",
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "product_id": {
                    "type": "integer",
                    "example": 3
                },
                "product_name": {
                    "type": "string",
                    "example": "Indomie Goreng"
                },
                "quantity": {
                    "type": "integer",
                    "example": 5
                },
                "subtotal": {
                    "type": "integer",
                    "example": 15000
                },
                "transaction_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}